+++ ReorderDatabase.Dtos/Note.cs
@@ -8,6 +8,6 @@
         // Ordering
         public ulong Numerator { get; set; }
         public ulong Denominator { get; set; }
-        public double Order { get; set; }
+        public decimal Order { get; set; }
     }
 }
